requirements.txt
Flask==2.2.2
flask-restful==0.3.9
python-dotenv==0.21.0
gunicorn==20.1.0
pytest==7.2.0
marqo==0.5.8